# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A e-commerce API server",
        "title": "ECOM API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
        "/admin/categories": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "創建新的商品類別(僅管理員)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "創建新類別",
                "parameters": [
                    {
                        "description": "類別資訊",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "創建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.CategoryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "請求參數錯誤",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    },
                    "401": {
                        "description": "未授權",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    },
                    "403": {
                        "description": "權限不足",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "伺服器錯誤",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    }
                }
            }
        },
        "/admin/categories/:id": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根據 id 更新商品類別(僅管理員)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "更新指定類別",
                "parameters": [
                    {
                        "description": "類別資訊",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "創建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.CategoryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "請求參數錯誤",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    },
                    "401": {
                        "description": "未授權",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    },
                    "403": {
                        "description": "權限不足",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "伺服器錯誤",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    }
                }
            }
        },
        "/admin/categories/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根據 ID 刪除商品類別(僅管理員)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "刪除指定類別",
                "parameters": [
                    {
                        "type": "string",
                        "description": "類別 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "刪除成功"
                    },
                    "400": {
                        "description": "請求參數錯誤",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    },
                    "401": {
                        "description": "未授權",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    },
                    "403": {
                        "description": "權限不足",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "伺服器錯誤",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    }
                }
            }
        },
...
        "/categories": {
            "get": {
                "description": "獲取所有類別列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "獲取所有類別",
                "responses": {
                    "200": {
                        "description": "獲取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.CategoryListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "伺服器錯誤",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    }
                }
            }
        },
        "/categories/{id}": {
            "get": {
                "description": "根據 ID 獲得類別詳情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "獲得單個類別",
                "parameters": [
                    {
                        "type": "string",
                        "description": "類別 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "獲取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.CategoryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "找不到",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "伺服器錯誤",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CategoryListResponse": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CategoryResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dto.CategoryResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.CreateCategoryRequest": {
            "type": "object",
            "required": [
                "name",
                "slug"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
...
        "dto.StandardResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.UpdateCategoryRequest": {
            "type": "object",
            "required": [
                "id",
                "name",
                "slug"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}